clean-arch: add a test for the proposed design pattern
